Fix issue with line parsing
Because of the way the data was casted into a list, there was a chance
that the is_on_ground value would be improperly parsed by the insert
statements.

This change removes the trailing characters and properly casts the value
as NULL if and when it is not sent with the ADSB message.

This prevents significant rollback and ensures that non-standard ADSB
messages are also captured.